Use uint16 as index type for OpenGl, split lines in vector tiles if resulting mesh is too long

+++ android/src/main/cpp/graphics/objects/LineGroup2dOpenGl.h
@@ -73,7 +73,7 @@ protected:
     GLuint vertexAttribBuffer = -1;
     std::vector<GLfloat> lineAttributes;
     GLuint indexBuffer = -1;
-    std::vector<GLuint> lineIndices;
+    std::vector<GLushort> lineIndices;
     bool glDataBuffersGenerated = false;
     Vec3D lineOrigin = Vec3D(0.0, 0.0, 0.0);
